Use unsigned long SampleTime in setter method
The SampleTime member variable was initially declared as int, but was
later changed to unsigned long in commit 169cdd0 (changed SampleTime
datatype, 2012-07-19).  However, the SetSampleTime() method's
signature remained unchanged and still awaits an int.

Update the signature of SetSampleTime() to accept an unsigned long
argument.

+++ PID_v1.cpp
@@ -102,7 +102,7 @@ void PID::SetTunings(double Kp, double Ki, double Kd)
 /* SetSampleTime(...) *********************************************************
  * sets the period, in Milliseconds, at which the calculation is performed	
  ******************************************************************************/
-void PID::SetSampleTime(int NewSampleTime)
+void PID::SetSampleTime(unsigned long NewSampleTime)
 {
    if (NewSampleTime > 0)
    {
@@ -110,7 +110,7 @@ void PID::SetSampleTime(int NewSampleTime)
                       / (double)SampleTime;
       ki *= ratio;
       kd /= ratio;
-      SampleTime = (unsigned long)NewSampleTime;
+      SampleTime = NewSampleTime;
    }
 }
